Moved _handleKeydown to Input feature.
