+++ src/input.js
@@ -5,6 +5,7 @@
 
 import Feature from '../core/feature.js';
 import InputCommand from './inputcommand.js';
+import { getCode } from '../utils/keyboard.js';
 
 /**
  * Handles text input coming from the keyboard or other input methods.
@@ -20,15 +21,77 @@ export default class Input extends Feature {
 		const editor = this.editor;
 		const editingView = editor.editing.view;
 
-		editor.commands.set( 'inputKeydown', new InputCommand( editor, 'keydown' ) );
-		editor.commands.set( 'inputMutation', new InputCommand( editor, 'mutation' ) );
+		editor.commands.set( 'inputMutation', new InputCommand( editor ) );
 
 		this.listenTo( editingView, 'keydown', ( evt, evtData ) => {
-			editor.execute( 'inputKeydown', { evtData } );
+			this._handleKeydown( evtData );
 		}, { priority: 'lowest' } );
 
 		this.listenTo( editingView, 'mutations', ( evt, mutations, viewSelection ) => {
 			editor.execute( 'inputMutation', { mutations, viewSelection } );
 		} );
 	}
+
+	/**
+	 * Handles the keydown event. We need to guess whether such keystroke is going to result
+	 * in typing. If so, then before character insertion happens, any selected content needs
+	 * to be deleted. Otherwise the default browser deletion mechanism would be
+	 * triggered, resulting in:
+	 *
+	 * * Hundreds of mutations which could not be handled.
+	 * * But most importantly, loss of control over how the content is being deleted.
+	 *
+	 * The method is used in a low-priority listener, hence allowing other listeners (e.g. input or enter features)
+	 * to handle the event.
+	 *
+	 * @private
+	 * @param {engine.view.observer.keyObserver.KeyEventData} evtData
+	 */
+	_handleKeydown( evtData ) {
+		const doc = this.editor.document;
+
+		if ( isSafeKeystroke( evtData ) || doc.selection.isCollapsed ) {
+			return;
+		}
+
+		doc.enqueueChanges( () => {
+			doc.composer.deleteContents( doc.batch(), doc.selection );
+		} );
+	}
+}
+
+const safeKeycodes = [
+	getCode( 'arrowUp' ),
+	getCode( 'arrowRight' ),
+	getCode( 'arrowDown' ),
+	getCode( 'arrowLeft' ),
+	16, // Shift
+	17, // Ctrl
+	18, // Alt
+	20, // CapsLock
+	27, // Escape
+	33, // PageUp
+	34, // PageDown
+	35, // Home
+	36, // End
+];
+
+// Function keys.
+for ( let code = 112; code <= 135; code++ ) {
+	safeKeycodes.push( code );
+}
+
+// Returns `true` if a keystroke should not cause any content change caused by "typing".
+//
+// Note: This implementation is very simple and will need to be refined with time.
+//
+// @param {engine.view.observer.keyObserver.KeyEventData} keyData
+// @returns {Boolean}
+function isSafeKeystroke( keyData ) {
+	// Keystrokes which contain Ctrl don't represent typing.
+	if ( keyData.ctrlKey ) {
+		return true;
+	}
+
+	return safeKeycodes.includes( keyData.keyCode );
 }
